requirements.txt
xgbse
xgboost
scikit-survival
lifelines 
pandas
joblib
fastapi
uvicorn
pydantic
numpy
scikit-learn
pytest
pytest-asyncio
httpx
pytest-mock
matplotlib
seaborn
apache-airflow==2.9.1
apache-airflow-providers-postgres
apache-airflow-providers-docker
apache-airflow-providers-slack
pyarrow